Added systemd check #86

+++ keyscore-agent/src/scripts/debian_pre_removal.sh
@@ -1,8 +1,13 @@
 #!/bin/bash
 
-# 1.1 Disable service
-systemctl disable ks_agent
-# 1.2 Stop service
-systemctl stop ks_agent
-#1.3 Remove service
-rm /etc/systemd/system/ks_agent.service
+#1 Check if systemd is available
+SYSTEMD="/usr/lib/systemd/"
+if [ -d $SYSTEMD ]
+  then
+    # 1.1 Disable service
+    systemctl disable ks_agent
+    # 1.2 Stop service
+    systemctl stop ks_agent
+    #1.3 Remove service
+    rm /etc/systemd/system/ks_agent.service
+fi
